pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "visionlabelops"
version = "0.1.0"
description = "Lightweight dataset audit, conversion, stats, split, report, and preview tooling for YOLO, COCO, and Labelme"
requires-python = ">=3.10"
readme = "README.md"
license = "MIT"
authors = [
  {name = "VisionLabelOps contributors"},
]
keywords = ["computer-vision", "dataset", "yolo", "coco", "labelme"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX :: Linux",
  "Operating System :: MacOS",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Image Recognition",
]
dependencies = [
  "Pillow>=10.0",
  "PyYAML>=6.0",
  "Jinja2>=3.1",
]

[project.scripts]
vlo = "visionlabelops.cli:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.6",
  "mypy>=1.10",
  "build>=1.2",
  "types-PyYAML>=6.0",
]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
visionlabelops = ["resources/*.j2", "py.typed"]

[tool.ruff]
target-version = "py310"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
no_implicit_optional = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
